+++ src/main/kotlin/com/setvect/bokslstock2/analysis/vbs/service/VbsBacktestService.kt
@@ -70,7 +70,7 @@ class VbsBacktestService(
                 val sellInfo = VbsTradeEntity(
                     vbsConditionEntity = condition,
                     tradeType = SELL,
-                    maPrice = currentCandle.average[condition.maPeriod] ?: 0,
+                    maPrice = currentCandle.average[condition.maPeriod] ?: 0.0,
                     yield = ApplicationUtil.getYield(lastBuyInfo.unitPrice, currentCandle.openPrice),
                     unitPrice = currentCandle.openPrice,
                     tradeDate = currentCandle.candleDateTimeStart
@@ -91,7 +91,7 @@ class VbsBacktestService(
                 lastBuyInfo = VbsTradeEntity(
                     vbsConditionEntity = condition,
                     tradeType = BUY,
-                    maPrice = currentCandle.average[condition.maPeriod] ?: 0,
+                    maPrice = currentCandle.average[condition.maPeriod] ?: 0.0,
                     yield = 0.0,
                     unitPrice = targetPrice,
                     tradeDate = currentCandle.candleDateTimeStart
@@ -104,10 +104,11 @@ class VbsBacktestService(
     /**
      * @return 목표가 계산
      */
-    private fun getTargetPrice(beforeCandle: CandleDto, currentCandle: CandleDto, condition: VbsConditionEntity): Int {
+    private fun getTargetPrice(beforeCandle: CandleDto, currentCandle: CandleDto, condition: VbsConditionEntity): Double {
         var volatilityPrice = (beforeCandle.highPrice - beforeCandle.lowPrice) * condition.kRate
         // 호가단위 기준으로 절삭
+        // TODO 브레이크포인트 걸어서 확인 필요
         volatilityPrice -= (volatilityPrice % condition.unitAskPrice)
-        return (currentCandle.openPrice + volatilityPrice).toInt()
+        return currentCandle.openPrice + volatilityPrice
     }
 }
